Initial recursion for local calls
